requirements.txt
numpy
colour-science
colour-checker-detection
matplotlib
jupyterlab
